refac: merge lib/api.js and lib/twitch/twitch.js in lib/twitch.js and refac
