+++ endpoints/main.js
@@ -1,22 +1,10 @@
-import { getChannelInfo, getTopStreamer, getUserByLogin } from '../lib/twitch/twitch.js'
+import { getTopStreamer } from '../lib/twitch.js'
 import { getStreamerLogin, getViewersFromLoginArray } from '../lib/twitch/helper.js'
 import { upsertHistory } from '../lib/twitch/db.js'
 
 export const routes = async (app, options) => {
     app.get('/', async function (request, reply) {
         reply.send({ data: 'main' })
-    })
-
-    app.get('/user/:login', async function (request, reply) {
-        let data = await getUserByLogin(request.params.login)
-
-        reply.send({ data })
-    })
-
-    app.get('/channel/:id', async function (request, reply) {
-        let data = await getChannelInfo(request.params.id)
-
-        reply.send({ data })
     })
 
     app.get('/streamer/:lang/:top', async function (request, reply) {
