+++ lib/twitch/helper.js
@@ -1,4 +1,4 @@
-import { getViewers } from './twitch.js'
+import { getViewers } from '../twitch.js'
 
 /**
  * Get all streamer login from a JSON containing data on streamers.
